# openapi.yaml
openapi: 3.0.3
info:
  title: NomosX API
  description: Agentic think tank API — research intelligence platform
  version: 2.0.0
  contact:
    name: NomosX Team
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://nomosx.netlify.app/api
    description: Production
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: nomosx-session
      description: JWT session cookie (set by /auth/login)
    AdminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin key for privileged endpoints
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /topics/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get a topic
      tags: [Topics]
      responses:
        '200': { description: Success }
        '404': { description: Not found }
    patch:
      summary: Update a topic
      tags: [Topics]
      security: [{ AdminKey: [] }]
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      summary: Delete a topic
      tags: [Topics]
      security: [{ AdminKey: [] }]
      responses:
        '200': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ============ BRIEFS ============
...
  /radar/subscribe:
    get:
      summary: List radar subscriptions
      tags: [Radar]
      security: [{ SessionCookie: [] }]
      responses:
        '200': { description: Subscriptions list }
    post:
      summary: Subscribe to radar topic
      tags: [Radar]
      security: [{ SessionCookie: [] }]
      responses:
        '201': { description: Subscribed }
    delete:
      summary: Unsubscribe from radar topic
      tags: [Radar]
      security: [{ SessionCookie: [] }]
      responses:
        '200': { description: Unsubscribed }

  # ============ SUBSCRIPTION ============
...